remove most platform specific code (#4)
* remove explicit support for building under DOS

or using specific DOS compilers

* remove more platform specific code

* remove obsolete build system support

* fix compilation in an MSYS2/UCRT64 environment

* run on Win64

* minor changes

+++ rott/audiolib/multivoc.h
@@ -105,7 +105,7 @@ int   MV_PlayVOC3D( char *ptr, int pitchoffset, int angle, int distance,
          int priority, unsigned long callbackval );
 int   MV_PlayVOC( char *ptr, int pitchoffset, int vol, int left, int right,
          int priority, unsigned long callbackval );
-int   MV_PlayLoopedVOC( char *ptr, long loopstart, long loopend,
+int   MV_PlayLoopedVOC( char *ptr, intptr_t loopstart, intptr_t loopend,
          int pitchoffset, int vol, int left, int right, int priority,
          unsigned long callbackval );
 void  MV_CreateVolumeTable( int index, int volume, int MaxVolume );
